requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.7
google-generativeai==0.3.0
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
pgvector==0.2.3
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.0

# Test dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
flake8==6.0.0
black==23.7.0
isort==5.12.0

# Development dependencies
requests==2.31.0